+++ draft-packages/badge/KaizenDraft/Badge/Badge.tsx
@@ -1,50 +1,52 @@
-import React, { useLayoutEffect, useState, useRef } from "react"
+import React, { HTMLAttributes, useLayoutEffect, useState } from "react"
 import classNames from "classnames"
-
+import { OverrideClassName } from "@kaizen/component-base"
 import styles from "./styles.module.scss"
 
-interface CommonProps {
-  readonly children?: string
+interface CommonProps
+  extends OverrideClassName<HTMLAttributes<HTMLSpanElement>> {
+  children?: string
   /**
    * The "dark" variant is no longer in the UI kit
    */
-  readonly variant?: "default" | "active" | "dark"
+  variant?: "default" | "active" | "dark"
   /**
    * renders reversed colors. Use on purple background
    */
-  readonly reversed?: boolean
+  reversed?: boolean
   /**
    * Supports "small" and "large" sizes - defaults to "small"
    */
-  readonly size?: "small" | "large"
+  size?: "small" | "large"
 }
-interface DotProps extends Omit<CommonProps, "children" | "variant"> {
-  readonly variant: "dot"
-  readonly children?: undefined
+
+interface DotProps extends Omit<CommonProps, "variant"> {
+  children?: undefined
+  variant: "dot"
 }
 
 export type BadgeProps = CommonProps | DotProps
 
-export const Badge = (props: BadgeProps) => {
-  const { children, variant = "default", reversed, size = "small" } = props
+export const Badge: React.VFC<BadgeProps> = ({
+  children,
+  variant = "default",
+  reversed = false,
+  size = "small",
+  classNameOverride,
+  ...restProps
+}) => (
+  <span
+    className={classNames(styles.badge, styles[variant], classNameOverride, {
+      [styles.reversed]: reversed,
+      [styles.large]: size === "large",
+    })}
+    {...restProps}
+  >
+    {variant !== "dot" && children}
+  </span>
+)
 
-  return (
-    <span
-      className={classNames(styles.badge, {
-        [styles.default]: variant === "default",
-        [styles.active]: variant === "active",
-        [styles.dark]: variant === "dark",
-        [styles.dot]: variant === "dot",
-        [styles.reversed]: reversed,
-        [styles.large]: size === "large",
-      })}
-    >
-      {variant != "dot" && children}
-    </span>
-  )
-}
-
-export const BadgeAnimated: React.FunctionComponent<BadgeProps> = props => {
+export const BadgeAnimated: React.VFC<BadgeProps> = props => {
   const [isFocused, setIsFocused] = useState(false)
 
   useLayoutEffect(() => {
@@ -64,4 +66,3 @@ export const BadgeAnimated: React.FunctionComponent<BadgeProps> = props => {
     </span>
   )
 }
-export default Badge