+++ packages/typography/src/Paragraph/Paragraph.tsx
@@ -32,7 +32,7 @@ export interface ParagraphProps
    * Not recommended. A short-circuit for overriding styles in a pinch
    * @default ""
    */
-  classNameAndIHaveSpokenToDST?: string
+  classNameOverride?: string
   children: React.ReactNode
   /**
    * HTML elements that are allowed on Paragraphs
@@ -47,7 +47,7 @@ export interface ParagraphProps
 }
 
 export const Paragraph = ({
-  classNameAndIHaveSpokenToDST,
+  classNameOverride,
   children,
   tag,
   variant,
@@ -58,7 +58,7 @@ export const Paragraph = ({
     styles.paragraph,
     styles[variant],
     styles[color],
-    classNameAndIHaveSpokenToDST,
+    classNameOverride,
   ])
 
   return createElement(