+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.12.11...@kaizen/draft-hero-card@1.12.12) (2022-03-17)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
+## [1.12.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.12.10...@kaizen/draft-hero-card@1.12.11) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [1.12.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.12.9...@kaizen/draft-hero-card@1.12.10) (2022-03-16)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card