+++ draft-packages/menu/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.5.18...@kaizen/draft-menu@4.5.19) (2022-03-17)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [4.5.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.5.17...@kaizen/draft-menu@4.5.18) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
+## [4.5.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.5.16...@kaizen/draft-menu@4.5.17) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/draft-menu
+
+
+
+
+
 ## [4.5.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu@4.5.15...@kaizen/draft-menu@4.5.16) (2022-03-16)
 
 **Note:** Version bump only for package @kaizen/draft-menu