+++ draft-packages/tag/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@2.3.17...@kaizen/draft-tag@2.3.18) (2022-03-17)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
+## [2.3.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@2.3.16...@kaizen/draft-tag@2.3.17) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
 ## [2.3.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@2.3.15...@kaizen/draft-tag@2.3.16) (2022-03-16)
 
 **Note:** Version bump only for package @kaizen/draft-tag