+++ packages/a11y/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.12...@kaizen/a11y@1.3.13) (2022-03-17)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
+## [1.3.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.11...@kaizen/a11y@1.3.12) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
 ## [1.3.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.10...@kaizen/a11y@1.3.11) (2022-03-16)
 
 **Note:** Version bump only for package @kaizen/a11y