+++ site/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-private.115](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.114...site@0.0.0-private.115) (2022-03-17)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.114](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.113...site@0.0.0-private.114) (2022-03-17)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.113](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.112...site@0.0.0-private.113) (2022-03-17)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.112](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.111...site@0.0.0-private.112) (2022-03-16)
+
+**Note:** Version bump only for package site
+
+
+
+
+
+# [0.0.0-private.111](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.110...site@0.0.0-private.111) (2022-03-16)
+
+
+### Features
+
+* rename `classNameAndIHaveSpokenToDST` to `classNameOverride` ([#2623](https://github.com/cultureamp/kaizen-design-system/issues/2623)) ([0ad2710](https://github.com/cultureamp/kaizen-design-system/commit/0ad2710f5e4b9a9d6b5a40ae72741a88669792c1))
+
+
+### BREAKING CHANGES
+
+* `classNameAndIHaveSpokenToDST` renamed to `classNameOverride`
+
+
+
+
+
+# [0.0.0-private.110](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.109...site@0.0.0-private.110) (2022-03-16)
+
+**Note:** Version bump only for package site
+
+
+
+
+
 # [0.0.0-private.109](https://github.com/cultureamp/kaizen-design-system/compare/site@0.0.0-private.108...site@0.0.0-private.109) (2022-03-16)